+++ app/models/zooms/app.rb
@@ -1,111 +1,51 @@
-class Zooms::App
-  def initialize(topic, start_time)
-    @topic = topic
-    @start_time = start_time
-  end
-
-  TIME_ZONE = 'Asia/Tokyo'.freeze
-
-  require 'net/http'
-  require 'base64'
-
-  def create!
-    response = post_api_request
-    body = JSON.parse(response.body)
-    raise StandardError, body['error'] unless response.code == '201'
-
-    body
-  end
-
-  private
-
-  def headers
-    token = create_access_token!
-    {
-      Authorization: "Bearer #{token}",
-      'Content-Type': 'application/json'
-    }
-  end
-
-  def post_api_request
-    uri = URI.parse(api_request_url)
-    http = Net::HTTP.new(uri.host, uri.port)
-    http.use_ssl = uri.scheme == 'https'
-
-    http.post(uri.path, request_body.to_json, headers)
-  end
-  
-  def create_access_token!
-    response = token_request
-
-    body = JSON.parse(response.body)
-    if response.code == '200'
-      body['access_token']
-    else
-      raise StandardError, body['error']
+module Zooms
+  class App
+    def initialize(topic, start_time)
+      @topic = topic
+      @start_time = start_time
+      @client = HttpClient.new
     end
-  end
 
-  def token_request
-    uri = URI.parse('https://zoom.us/oauth/token')
-    http = Net::HTTP.new(uri.host, uri.port)
-    http.use_ssl = uri.scheme === 'https'
+    TIME_ZONE = 'Asia/Tokyo'.freeze
+    OWNER_EMAIL = Rails.application.credentials.api[:owner_email]
 
-    http.post(uri.path, token_request_query.to_param, token_request_headers)
-  end
+    def create!
+      token = @client.create_access_token!
+      response = @client.post(api_request_url, request_body, headers(token))
+      body = JSON.parse(response.body)
+      raise StandardError, body['error'] unless response.code == '201'
 
-  def token_request_query
-    {
-      grant_type: 'account_credentials',
-      account_id:,
-    }
-  end
+      body
+    end
 
-  def token_request_headers
-    {
-      'authorization' => "Basic #{credentials}",
-      'content-type' => 'application/x-www-form-urlencoded'
-    }
-  end
+    private
 
-  def api_request_url
-    "https://api.zoom.us/v2/users/#{owner_email}/meetings"
-  end
+    def headers(token)
+      {
+        Authorization: "Bearer #{token}",
+        'Content-Type': 'application/json'
+      }
+    end
 
-  def request_body
-    # https://marketplace.zoom.us/docs/api-reference/zoom-api/methods/#operation/meetingCreate
-    {
-      topic: @topic,
-      type: '2',
-      start_time: @start_time, # yyyy-MM-ddTHH:mm:ss
-      timezone: TIME_ZONE,
-      duration: 30, # 所要時間
-      password:,
-    }
-  end
+    def api_request_url
+      "https://api.zoom.us/v2/users/#{OWNER_EMAIL}/meetings"
+    end
 
-  def password
-    random = Random.new
-    random.rand(9_999_999_999)
-  end
+    def request_body
+      # https://marketplace.zoom.us/docs/api-reference/zoom-api/methods/#operation/meetingCreate
+      {
+        topic: @topic,
+        type: '2',
+        start_time: @start_time, # yyyy-MM-ddTHH:mm:ss
+        timezone: TIME_ZONE,
+        duration: 30, # 所要時間
+        password: password
+      }
+    end
 
-  def credentials
-    Base64.strict_encode64("#{client_id}:#{client_secret}")
-  end
-
-  def account_id
-    Rails.application.credentials.api[:account_id]
-  end
-
-  def client_id
-    Rails.application.credentials.api[:client_id]
-  end
-
-  def client_secret
-    Rails.application.credentials.api[:client_secret]
-  end
-
-  def owner_email
-    Rails.application.credentials.api[:owner_email]
+    def password
+      random = Random.new
+      random.rand(9_999_999_999)
+    end
   end
 end
